Changed the type of player IDs from byte to int
In bigger servers, bytes really don't cut it. Shorts could work, but might as well go for the full 32-bits at this point. The only reason why I didn't go for long is that it's really overkill at that point. Also, Random.nextLong can't generate every possible values and I'm too lazy to code a workaround.

+++ src/net/fe/FEMultiplayer.java
@@ -421,7 +421,7 @@ public class FEMultiplayer extends Game{
 	 *
 	 * @return the players
 	 */
-	public static HashMap<Byte, Player> getPlayers() {
+	public static HashMap<Integer, Player> getPlayers() {
 		return getSession().getPlayerMap();
 	}
